addd snapshot tests
